Moved tag editor out of view component so only one is created
