+++ src/app/container/versions/versions.component.html
@@ -101,6 +101,7 @@
       </tr>
     </tbody>
   </table>
+  <app-version-modal></app-version-modal>
 </div>
 <div class="row pull-right">
   <app-add-tag *ngIf="!publicPage"></app-add-tag>
