+++ src/app/workflow/versions/versions.component.html
@@ -50,7 +50,6 @@
           Actions
         </label>
       </th>
-      <app-version-modal></app-version-modal>
     </tr>
     </thead>
     <tbody>
@@ -81,4 +80,5 @@
     </tr>
     </tbody>
   </table>
+  <app-version-modal></app-version-modal>
 </div>
